requirements.txt
selenium 
webdriver-manager 
pandas
bs4
requests
pinecone
pinecone-client 
langchain-openai 
langchain-pinecone 
pandas 
python-dotenv
langchain_huggingface
huggingface-hub
sentence-transformers
langchain
langchain-core
langchain-google-genai
fastapi
uvicorn